Added more if/else checks to figure out how to format rows properly. Rows with numbers and no dashes will have separate <td> tags per token excluding the first instances with no numbers.
Dashes will work the same way as numbers in terms of <td> tagging. Code
is almost exactly the same.

Added spacing for easier reading. Mainly for place names in which
appended string was all joined together.

Still need a check for the Valid to/from date in which is currently in a
table cell itself when it shouldn't be. This warps the format slightly
but isn't a major problem. Should still be fixed though.