Fixed so Hillhouse farm rd is now 1 word and rd is not split into a separate cell (wrong variable name used in containsDash if method - should be dashposition and not position).
Added two new ifs. One to check for the word "valid" which should have
more than 6 digits for the date but should not be formatted into
separate cells.

Second if, looks for rows with both dashes and numbers. Finds the first
occurence of both and then compares to find the lower position and using
this position, will format the row appropriately - similarly to the
numbers and dashes if previously. Just it's tailored for these annoying
rows that have both.