Numbered anchors finally added and working. First term is initialized to 1. With < and > to indicate jump to next term. Eg. First term is Hamilton1, < is a link to Hamilton0 and > is a link to Hamilton2. Term is displayed as "Hamilton" with the number "Hamilton2" etc, being inside the a href link and span id tags for anchor links.
This is done by splitting up the final formatted contents via </tr> so
contents are returned in a single row. Each row is then checked for the
geoterm and replaced with the newly formatted HTML string for anchor
links. Sadly this only replaces the FIRST instance of the term per row
but is not likely to cause a major issue as any duplicates in the same
row is often easily read without the need for an "extra jump".
Unfortunately this also means it is not highlighted.