Major bugfix: For addresses that are not reverse geocoded properly in which the separate components are returned as NULL. The full address is now split and highlighting + jump links are included in relation to the tokenized address.
The issue arises when the address contains stopwords or duplicate terms.
For example "on" must be removed as the rows use the contain method to
check for geoterms instead of checking every single token of the row
when highlighting. Therefore it will end up highlighting every word that
contains "on". Issue 2: when breaking the address into separate tokens,
duplicates are often introduced. Eg. Glasgow, Glasgow City --> Glasgow,
Glasgow, City. The two Glasgows are then included in the highlight loop
and causes some severe HTML errors (double <span id> tags) which ruins
the entire formatting. 

This is now fixed by converting the list to a hashset and back to remove
the duplicates.

A check must still be implemented (possibly if necessary) in which the
address break up is only used IF the geocoder returns Null results (also
need to look into why this happens for some specific places).